requirements.txt
# Main Librairies
pandas==2.3.2
dotenv==0.9.9

# Vusualization
matplotlib==3.10.5
seaborn==0.13.2

# NLP
nltk==3.9.1
wordcloud==1.9.4
gensim==4.3.3
pyspellchecker==0.8.3

# Machine Learning
scikit-learn==1.7.1
ipywidgets==8.1.7
mlflow==3.3.2
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
ipykernel

transformers==4.57.0
torch==2.8.0

streamlit==1.50.0
plotly==6.3.1
# Add FastAPI and related packages
fastapi==0.110.0
uvicorn==0.30.0
pydantic==2.7.0
requests==2.32.3

# Unit testing
pytest==8.4.2
